docs: refactor stories path
