+++ .storybook/main.ts
@@ -2,13 +2,9 @@ export default {
   framework: {
     name: '@storybook/react-vite',
   },
-  core: {
-    builder: '@storybook/builder-vite',
-  },
   stories: [
     '../packages/react-core/**/*.stories.@(tsx|jsx)',
     '../packages/react-web-kit/**/*.stories.@(tsx|jsx)',
-    '../packages/cortex-core/**/*.stories.@(tsx|jsx)',
     '../packages/cortex-react/**/*.stories.@(tsx|jsx)',
     '../docs/**/*.mdx',
   ],
