+++ packages/cortex-react/docs/input.stories.tsx
@@ -1,11 +1,10 @@
 import React from 'react';
 import { StoryFn } from '@storybook/react';
-import { Input } from '../components';
-import { IoPerson } from 'react-icons/io5';
-import { IoEye } from 'react-icons/io5';
+import { Input } from '../src/components';
+
 export default {
-  title: 'Cortex-React/Input/Custom',
-  component: <div />,
+  title: 'Cortex/Input',
+  component: Input,
   args: {
     label: 'Label',
     placeholder: 'Placeholder',
@@ -27,15 +26,11 @@ export default {
 
 const Template: StoryFn = args => {
   return (
-    <Input.Face variants={{ intent: args.intent }}>
-      <Input.Left>
-        <IoPerson size={16} />
-      </Input.Left>
-      <Input.Box label={args.label} placeholder={args.placeholder}></Input.Box>
-      <Input.Right>
-        <IoEye />
-      </Input.Right>
-    </Input.Face>
+    <Input.Root
+      variants={{ intent: args.intent }}
+      label={args.label}
+      placeholder={args.placeholder}
+    />
   );
 };
