+++ packages/cortex-react/docs/tag.stories.tsx
@@ -1,10 +1,10 @@
 import React from 'react';
 import { StoryFn } from '@storybook/react';
-import { Tag } from '../index';
+import { Tag } from '../src';
 
 export default {
-  title: 'Cortex-React/Tag',
-  component: <div />,
+  title: 'Cortex/Tag',
+  component: Tag,
   args: {
     intent: 'primary',
   },
@@ -20,13 +20,6 @@ const Template: StoryFn = args => {
   return (
     <div className={'flex flex-col justify-center items-center'}>
       <Tag variants={{ intent: args.intent }} label={'Intent color'} />
-      <div className={'flex flex-row gap-x-deca mt-giga'}>
-        <Tag variants={{ intent: 'primary' }} label={'Tag label'} />
-        <Tag variants={{ intent: 'success' }} label={'Tag label'} />
-        <Tag variants={{ intent: 'info' }} label={'Tag label'} />
-        <Tag variants={{ intent: 'secondary' }} label={'Tag label'} />
-        <Tag variants={{ intent: 'white' }} label={'Tag label'} />
-      </div>
     </div>
   );
 };
